Fix width argument, set overall width (#825)
Fix the width for the description details, both default and as cmdline option,
and make sure we don't break the width of the calw and calm output.
The (pretty much undocumented) option --width/-w option was used for setting the
width of a day table cell in calw and calm but seemed to also be intended for
setting the overall width of the description in agenda. The latter part was
broken and the details description width would always default to 80.
We change the behaviour to set the overall width of the table, defaulting to the
terminal width, both for calw and calm and for description details. The width of
the day cell is calculated from the overall width.